+++ steam_web/Models/AppContextData.cs
@@ -16,7 +16,7 @@ public class AppContextData
     [JsonPropertyName("trade_permissions")] public string TradePermissions { get; init; }
     [JsonPropertyName("load_failed")] public byte LoadFailed { get; init; } = 0;
     [JsonPropertyName("owner_only")] public bool OwnerOnly { get; init; } = false;
-    [JsonPropertyName("rgContexts")] public Dictionary<byte, ContextItem> rgContexts { get; init; } = new(3);
+    [JsonPropertyName("rgContexts")] public Dictionary<uint, ContextItem> rgContexts { get; init; } = new(3);
 
     public static Dictionary<uint, AppContextData> Deserialize(string data)
     {
@@ -30,6 +30,6 @@ public class AppContextData
 public class ContextItem
 {
     [JsonPropertyName("asset_count")] public ushort AssetsCount { get; init; }
-    [JsonPropertyName("id")] public byte ContextId { get; init; }
+    [JsonPropertyName("id")] public uint ContextId { get; init; }
     [JsonPropertyName("name")] public string Name { get; init; }
 }
